tpsの小数対応

+++ src/main/java/ramune314159265/wsconnectionplugin/TickRunnable.java
@@ -3,13 +3,13 @@ package ramune314159265.wsconnectionplugin;
 import java.util.Date;
 
 public class TickRunnable implements Runnable{
-	public static int Tps = 20;
+	public static float Tps = 20;
 	public static long lastTickTimestamp = new Date().getTime();
 	public long lastTickTime= System.currentTimeMillis();
 
 	public void run(){
 		long lastTickElapsed = System.currentTimeMillis() - lastTickTime;
-		TickRunnable.Tps = (int) (1000 / lastTickElapsed);
+		TickRunnable.Tps = (float) (1000 / lastTickElapsed);
 		lastTickTime = System.currentTimeMillis();
 		TickRunnable.lastTickTimestamp = new Date().getTime();
 	}
