+++ src/main/java/ramune314159265/wsconnectionplugin/events/EverySecond.java
@@ -4,9 +4,9 @@ import ramune314159265.wsconnectionplugin.WsConnectionPlugin;
 
 public class EverySecond extends Event{
 	public String serverId;
-	public int tps;
+	public float tps;
 	public long lastTickTimestamp;
-	public EverySecond(int tps, long lastTickTimestamp){
+	public EverySecond(float tps, long lastTickTimestamp){
 		this.type = "every_second_info_send";
 		this.tps = tps;
 		this.lastTickTimestamp = lastTickTimestamp;
